add zbuffer

+++ src/include/renderer.h
@@ -42,7 +42,7 @@ typedef struct context {
     vec3 pos;
     vec3 rot; // rot around x, y, z axes respectively
     point *proj; // projected vertices in last frame (also includes depth as z)
-    float *zbuf; // depth buffer
+    uint32_t *zbuf; // depth buffer
     double flength; // focal length
     double brightness;
     SDL_Renderer *renderer;
